add scoped require.
